Clean up nanobind bindings and Python type hints; fix C++ API docs (#1048)
* Fix inaccuracies and improve clarity in C++ API documentation

Corrects Doxygen comments that contradicted the implementation
(mismatched parameter order/shapes, stale/copy-pasted @return text,
wrong in/out tags, an inverted-shape description, and a namespace brief
that was too narrow), adds missing @param docs to the Piola map
functions and several precompute helpers, documents element::family
enum values, and fixes assorted typos and grammar issues.


* Fix Doxygen build failure from @param tags on unnamed parameters

maps.h's Piola-map functions leave J/detJ/K unnamed (commented out)
where unused, to suppress unused-parameter warnings. Doxygen's
WARN_AS_ERROR build failed because @param can't match a tag to a
parameter that has no name in the signature. Replace the offending
@param entries with a prose note on which inputs each map ignores.


* Fix nanobind array-mutation bug and clean up Python type hints

Add .noconvert() to the mutable ndarray arguments of T_apply,
Tt_apply_right, Tt_inv_apply, permute_subentity_closure and
permute_subentity_closure_inv. Without it, nanobind silently created a
converted temporary when a caller passed a mismatched dtype/layout
array, mutated the throwaway temporary in place, and left the caller's
array untouched with no error. This had made
test_dof_transformations.py::test_apply_right vacuous (it fed float32
arrays into float64 elements), so its arrays are switched to the
matching dtype.

Also remove `# type: ignore` comments that are no longer needed now
that nanobind's generated stubs type these APIs correctly (confirmed
via mypy --warn-unused-ignores), and fix a few underlying type
inaccuracies instead of suppressing them (an ArrayLike-vs-NDArray
property type in ufl.py, a float64-only helper in polynomials.py, and
incorrect entity_dofs annotations in numba_helpers.py). Remaining
irreducible dtype-dispatch cases use scoped cast() instead of blanket
ignores.


* Name arguments in remaining nanobind bindings

Add nb::arg names to the ~30 bindings in wrapper.cpp that previously
had none, so the generated _basixcpp.pyi stubs show meaningful
parameter names (e.g. sub_entity_type(celltype, dim, index)) instead
of positional-only arg0/arg1/arg2. Names are chosen to match whatever
the corresponding public basix.* Python wrapper calls each argument,
including the tabulate_polynomial_set_* labels which previously didn't
match their caller (polytype/d/n/x vs. the wrapper's ptype/degree/
nderiv/pts). These parameters are now keyword-callable from Python in
addition to positional.


* Add missing type hints and collapse dtype-dispatch duplication

finite_element.py: push_forward, T_apply, Tt_apply_right and
Tt_inv_apply had no parameter type hints at all, unlike the rest of
the class. Typing data as npt.NDArray[T] now correctly documents that
it must match the element's own dtype (the constraint the earlier
.noconvert() fix enforces at runtime); resolve the resulting
generic-dispatch mismatch the same way as tabulate, via cast(Any, ...).

ufl.py: _ElementBase.get_component_element's abstract declaration
returned tuple[_typing.Any, int, int] only because _ElementBase isn't
bound yet inside its own class body; every concrete override already
returns tuple[_ElementBase, int, int]. Use a quoted forward reference
instead.

wrapper.cpp: collapse the identical dtype-dispatch if/else-throw block
duplicated across create_element, create_tp_element and tp_factors
into a single dispatch_dtype<R>(dtype, f) helper using a C++20
templated lambda, so each binding only states its own call.


* Tighten entity_transformations return type

entity_transformations() was typed as a bare dict while every sibling
method on FiniteElement is precisely typed. Type it as
dict[str, npt.NDArray[T]], matching what the underlying binding
actually returns (cell-type-name keys, per-T ndarray values).


* Enable mypy testing for Windows

Add a mypy step to both Windows CI jobs (combined and split builds),
mirroring the check already run in the no-isolation-python-build job
on Linux. Both jobs already install mypy (and, transitively via the
optional extra, fenics-ufl) through the .[ci] extra, so no new
dependency-install steps are needed.

* Ship py.typed reliably on the Windows split build

nanobind_add_stub's INSTALL_TIME MARKER_FILE generation has to import
basix._basixcpp to introspect it. On the split Windows build that
import can fail (DLL load failed - libbasix.dll isn't discoverable
yet at that point in the install, per the existing comment above),
which silently skips marker-file creation. Confirmed via CI logs
(run 30583257866): the combined-build job writes the marker
successfully and mypy is clean, while the split-build job never
writes it and mypy reports "Package 'basix' cannot be type checked
due to missing py.typed marker."

Install the static py.typed marker directly, matching the reliable
mechanism the non-Windows branch already uses, so the package is
always recognised as typed regardless of whether the DLL-import-
dependent stub generation succeeds.


---------

+++ test/test_dof_transformations.py
@@ -50,7 +50,7 @@ def test_apply_right(cell_type, element_type, degree, element_args):
     random.seed(42)
     e = cached_create_element(element_type, cell_type, degree, tuple(element_args))
     size = e.dim
-    base1 = np.arange(size * size, dtype=np.float32)
+    base1 = np.arange(size * size, dtype=np.float64)
     # This is the transpose of base1
     base2 = base1.reshape(size, size).T.reshape(-1).copy()
     for i in range(10):
